Added a standalone feature in section-events (#5)
to define sections containing either FailureReporte or ConfigurationReport as a standalone. The section is then treated as a standalone one which means that it won't be registered in the section tree and the attached report won't be merged with any existing one, so it will be added into the list of configuration/failure reports, no matter what type it is.